gen3sis2: General Engine for Eco-Evolutionary Simulations

Contains an engine for spatially-explicit eco-evolutionary mechanistic models with a modular implementation and several support functions. It allows exploring the consequences of ecological and macroevolutionary processes across realistic or theoretical spatio-temporal spaces on biodiversity patterns as a general term. Reference: Oskar Hagen, Benjamin Flueck, Fabian Fopp, Juliano S. Cabral, Florian Hartig, Mikael Pontarp, Thiago F. Rangel, Loic Pellissier (2021) "gen3sis: A general engine for eco-evolutionary simulations of the processes that shape Earth's biodiversity" doi:10.1371/journal.pbio.3001340 .

See also

Author

Maintainer: Oskar Hagen oskar@hagen.bio (space Ecology, WSL and ETH Zürich, Switzerland)

Authors:

Oskar Hagen oskar@hagen.bio (space Ecology, WSL and ETH Zürich, Switzerland)

Admir C. de O. Junior admircjunior@gmail.com

Bouwe Reijenga bouwe.reijenga@earth.ox.ac.uk

Benjamin Flück benjamin.flueck@alumni.ethz.ch (space Ecology, WSL and ETH Zürich, Switzerland)

Fabian Fopp fabian.fopp@usys.ethz.ch (space Ecology, WSL and ETH Zürich, Switzerland)

Juliano S. Cabral juliano.sarmento_cabral@uni-wuerzburg.de (Ecosystem Modeling, Center for Computational and Theoretical Biology, University of Würzburg, Würzburg, Germany)

Florian Hartig florian.hartig@biologie.uni-regensburg.de (Theoretical Ecology, University of Regensburg, Regensburg, Germany)

Mikael Pontarp mikael.pontarp@biol.lu.se (Department of Biology, Lund University, Lund, Sweden)

Thiago F. Rangel rangel.tf@gmail.com (Department of Ecology, Universidade Federal de Goiás, Goiás, Brazil)

Loïc Pellissier loic.pellissier@usys.ethz.ch (space Ecology, WSL and ETH Zürich, Switzerland) [thesis advisor]

Other contributors:

ETH Zürich [copyright holder]

Charles Novaes de Santana charles.novaes@usys.ethz.ch (space Ecology, WSL and ETH Zürich, Switzerland) [contributor]

Theo Gaboriau theo.gaboriau@unil.ch (Department of Computational Biology, Lausanne University, Switzerland) [contributor]